Minor shared preferences refactoring and few string changes
