+++ app/src/test/java/org/openedx/AppViewModelTest.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LifecycleRegistry
 import org.openedx.app.data.storage.PreferencesManager
-import org.openedx.core.domain.model.User
+import org.openedx.core.data.model.User
 import org.openedx.app.room.AppDatabase
 import org.openedx.app.system.notifier.AppNotifier
 import org.openedx.app.system.notifier.LogoutEvent
