+++ core/src/main/java/org/openedx/core/data/storage/CorePreferences.kt
@@ -1,6 +1,6 @@
 package org.openedx.core.data.storage
 
-import org.openedx.core.domain.model.User
+import org.openedx.core.data.model.User
 import org.openedx.core.domain.model.VideoSettings
 
 interface CorePreferences {
