+++ core/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name" translatable="false">OpenEdX</string>
+    <string name="app_name" translatable="false">@string/platform_name</string>
 
     <string name="core_results">Results</string>
     <string name="core_error_invalid_grant">Invalid credentials</string>
@@ -31,7 +31,7 @@
     <string name="video_quality_p540">540p</string>
     <string name="video_quality_p720">720p (Best quality)</string>
     <string name="core_user_not_active">User account is not activated. Please activate your account first.</string>
-    <string name="core_email_chooser_header">Send email using...</string>
+    <string name="core_email_chooser_header">Send email using…</string>
     <string name="core_email_client_not_present">No e-mail clients installed</string>
     <string name="core_app_version">App Version: </string>
     <string name="core_android_os_version">Android OS: </string>
