+++ profile/src/main/java/org/openedx/profile/data/storage/ProfilePreferences.kt
@@ -1,6 +1,6 @@
 package org.openedx.profile.data.storage
 
-import org.openedx.profile.domain.model.Account
+import org.openedx.profile.data.model.Account
 
 interface ProfilePreferences {
     var profile: Account?
